Fixing Asset Structure
Hello!
This is the skeleton for the research project pages. Due to a bug with images + data src, we had to restructure what we considered what a common field was. Each entry now has their own featureImage field. The matrix details was moved into research.

We nuked the asset folder to start fresh.

The bug still exists. Sad.